+++ annotation-processor/pom.xml
@@ -82,13 +82,13 @@
 		<dependency>
 			<groupId>org.testcontainers</groupId>
 			<artifactId>testcontainers</artifactId>
-			<version>1.14.3</version>
+			<version>1.15.0-rc2</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.testcontainers</groupId>
 			<artifactId>mysql</artifactId>
-			<version>1.14.3</version>
+			<version>1.15.0-rc2</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
