+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/AnnotationProcessorApplication.java
@@ -1,0 +1,15 @@
+package com.frontier.api.annotation.processor;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AnnotationProcessorApplication {
+
+  public AnnotationProcessorApplication() {
+  }
+
+  public static void main(String[] args) {
+    SpringApplication.run(AnnotationProcessorApplication.class, args);
+  }
+}
