+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/annotation/consumer/FrontierConsumerRepository.java
@@ -1,0 +1,17 @@
+package com.frontier.api.annotation.processor.annotation.consumer;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+@Documented
+public @interface FrontierConsumerRepository {
+
+  String guarantee();
+
+}
+
