+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/annotation/provider/repository/FrontierProviderRepository.java
@@ -1,0 +1,17 @@
+package com.frontier.api.annotation.processor.annotation.provider.repository;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.stereotype.Repository;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE})
+@Documented
+@Repository
+public @interface FrontierProviderRepository {
+
+}
+
