+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/immutables/api/FrontierAPIInterface.java
@@ -1,0 +1,10 @@
+package com.frontier.api.annotation.processor.immutables.api;
+
+import java.util.Set;
+
+public interface FrontierAPIInterface<T> {
+
+  T produceMessage(FrontierApiIdentity frontierApiIdentity, Set<Object> methodParams);
+
+}
+
