+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/immutables/api/FrontierApiNode.java
@@ -1,0 +1,17 @@
+package com.frontier.api.annotation.processor.immutables.api;
+
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+@Builder
+@Getter
+@EqualsAndHashCode
+public class FrontierApiNode {
+
+  private String serviceName;
+
+  private String beanName;
+
+  private String methodName;
+}
