+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/immutables/domain/FrontierRepositoryProperty.java
@@ -1,0 +1,14 @@
+package com.frontier.api.annotation.processor.immutables.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class FrontierRepositoryProperty {
+
+  private Guarantee guarantee;
+
+  private String methodName;
+
+}
