+++ annotation-processor/src/main/java/com/frontier/api/annotation/processor/service/FrontierRepositoryWrapperService.java
@@ -1,0 +1,55 @@
+package com.frontier.api.annotation.processor.service;
+
+import com.frontier.api.annotation.processor.immutables.domain.FrontierRepositoryIdentity;
+import com.frontier.api.annotation.processor.immutables.domain.FrontierRepositoryProperty;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+
+public class FrontierRepositoryWrapperService {
+
+  private final Map<FrontierRepositoryIdentity, Set<FrontierRepositoryProperty>> frontierRepositoryProperties =
+      new HashMap<>();
+
+  private static FrontierRepositoryWrapperService frontierRepositoryWrapperServiceInstance;
+
+  public static FrontierRepositoryWrapperService getInstance() {
+    if (frontierRepositoryWrapperServiceInstance == null) {
+      frontierRepositoryWrapperServiceInstance = new FrontierRepositoryWrapperService();
+    }
+    return frontierRepositoryWrapperServiceInstance;
+  }
+
+  private FrontierRepositoryWrapperService() {
+  }
+
+  public void addFrontierRepositoryProperty(FrontierRepositoryIdentity frontierRepositoryIdentity,
+      FrontierRepositoryProperty frontierRepositoryProperty) {
+    Set<FrontierRepositoryProperty> updatedFrontierProperties = Optional
+        .ofNullable(this.frontierRepositoryProperties.get(frontierRepositoryIdentity))
+        .map(f -> {
+          f.add(frontierRepositoryProperty);
+          return f;
+        })
+        .orElseGet(() -> {
+          HashSet<FrontierRepositoryProperty> frontierProperties = new HashSet<>();
+          frontierProperties.add(frontierRepositoryProperty);
+          return frontierProperties;
+        });
+    this.frontierRepositoryProperties.put(frontierRepositoryIdentity, updatedFrontierProperties);
+  }
+
+  public void addFrontierRepositoryIdentity(FrontierRepositoryIdentity frontierRepositoryIdentity) {
+    Set<FrontierRepositoryProperty> emptyFrontierProperties = Optional
+        .ofNullable(this.frontierRepositoryProperties.get(frontierRepositoryIdentity))
+        .orElseGet(HashSet::new);
+    this.frontierRepositoryProperties.put(frontierRepositoryIdentity, emptyFrontierProperties);
+  }
+
+  public Map<FrontierRepositoryIdentity, Set<FrontierRepositoryProperty>> getFrontierRepositoryProperties() {
+    return this.frontierRepositoryProperties;
+  }
+
+}
