--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/AnnotationProcessorApplication.java
@@ -1,15 +1,0 @@
-package com.frontier.api.annotationprocessor;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AnnotationProcessorApplication {
-
-  public AnnotationProcessorApplication() {
-  }
-
-  public static void main(String[] args) {
-    SpringApplication.run(AnnotationProcessorApplication.class, args);
-  }
-}
