--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/api/FrontierApiIdentity.java
@@ -1,15 +1,0 @@
-package com.frontier.api.annotationprocessor.api;
-
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Builder
-@Getter
-@EqualsAndHashCode
-public class FrontierApiIdentity {
-
-  private String beanName;
-
-  private String methodName;
-}
