--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/api/FrontierApiRegisterResponseMessage.java
@@ -1,17 +1,0 @@
-package com.frontier.api.annotationprocessor.api;
-
-import java.util.List;
-import java.util.Map;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Builder
-@Getter
-@EqualsAndHashCode
-public class FrontierApiRegisterResponseMessage {
-
-  private Map<String, List<FrontierApiIdentity>> frontierIdentitiesByServiceName;
-
-}
-
