--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/domain/FrontierRepositoryIdentity.java
@@ -1,15 +1,0 @@
-package com.frontier.api.annotationprocessor.domain;
-
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Builder
-@Getter
-@EqualsAndHashCode
-public class FrontierRepositoryIdentity {
-
-  private String classpath;
-
-  private String beanName;
-}
