--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/domain/FrontierRepositoryProperty.java
@@ -1,14 +1,0 @@
-package com.frontier.api.annotationprocessor.domain;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class FrontierRepositoryProperty {
-
-  private Guarantee guarantee;
-
-  private String methodName;
-
-}
