--- annotation-processor/src/main/java/com/frontier/api/annotationprocessor/domain/Guarantee.java
@@ -1,14 +1,0 @@
-package com.frontier.api.annotationprocessor.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum Guarantee {
-  SYNCHRONOUS("synchronous"),
-  ASYNCHRONOUS("asynchronous"),
-  BEST_EFFORT("best-effort");
-
-  private String name;
-}
