--- annotation-processor/src/test/java/com/frontier/api/annotationprocessor/test/User.java
@@ -1,96 +1,0 @@
-package com.frontier.api.annotationprocessor.test;
-
-import java.util.Objects;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class User {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-  private String firstName;
-  private String lastName;
-  private String email;
-
-  public User() {
-  }
-
-  public User(String firstName, String lastName, String email) {
-    this.firstName = firstName;
-    this.lastName = lastName;
-    this.email = email;
-  }
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public void setFirstName(String firstName) {
-
-    this.firstName = firstName;
-  }
-
-  public String getLastName() {
-
-    return lastName;
-  }
-
-  public void setLastName(String lastName) {
-
-    this.lastName = lastName;
-  }
-
-  public String getEmail() {
-
-    return email;
-  }
-
-  public void setEmail(String email) {
-
-    this.email = email;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    User user = (User) o;
-    return Objects.equals(id, user.id) &&
-        Objects.equals(firstName, user.firstName) &&
-        Objects.equals(lastName, user.lastName) &&
-        Objects.equals(email, user.email);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, firstName, lastName, email);
-  }
-
-  @Override
-  public String toString() {
-    final StringBuilder sb = new StringBuilder("User{");
-    sb.append("id=").append(id);
-    sb.append(", firstName='").append(firstName).append('\'');
-    sb.append(", lastName='").append(lastName).append('\'');
-    sb.append(", email='").append(email).append('\'');
-    sb.append('}');
-    return sb.toString();
-  }
-}
-
